+++ packages/credential-w3c/src/__tests__/message-handler.test.ts
@@ -26,7 +26,7 @@ describe('@veramo/credential-w3c', () => {
             didResolutionMetadata: {},
             didDocumentMetadata: {},
             didDocument: {
-              '@context': 'https://w3id.org/did/v1',
+              '@context': 'https://www.w3.org/ns/did/v1',
               id: args?.didUrl,
               verificationMethod: [
                 {