+++ packages/did-comm/src/__tests__/packing.test.ts
@@ -47,7 +47,7 @@ const multiBaseDoc = {
     {
       "controller": "did:web:portcullis.1keep.com",
       "id": "did:web:portcullis.1keep.com#1",
-      "publicKeyMultibase": "z7bNjXnt3EXhpy7UdzZUANa2CsBFpzKGFXsgvkBd9iHGo",
+      "publicKeyMultibase": "z6LSpSrLxbAhg2SHwKk7kwpsH7DM7QjFS5iK6qP87eViohud",
       "type": "X25519KeyAgreementKey2020"
     }
   ]
@@ -94,7 +94,7 @@ const base58Doc = {
 }
 
 const hexDoc = {
-  "@context": "https://w3id.org/did/v1",
+  "@context": "https://www.w3.org/ns/did/v1",
   "id": "did:web:iiw-demo.herokuapp.com",
   "verificationMethod": [
     {