+++ packages/remote-server/src/web-did-doc-router.ts
@@ -1,6 +1,7 @@
-import { IIdentifier, IDIDManager, TAgent, TKeyType } from '@veramo/core-types'
+import { IIdentifier, IDIDManager, TAgent, TKeyType, IKey } from '@veramo/core-types'
+import { bytesToBase58, bytesToMultibase, hexToBytes } from '@veramo/utils'
 import { Request, Router } from 'express'
-import { ServiceEndpoint } from 'did-resolver'
+import { ServiceEndpoint, VerificationMethod } from 'did-resolver'
 
 interface RequestWithAgentDIDManager extends Request {
   agent?: TAgent<IDIDManager>
@@ -41,12 +42,52 @@ export const WebDidDocRouter = (options: WebDidDocRouterOptions): Router => {
   const router = Router()
 
   const didDocForIdentifier = (identifier: IIdentifier) => {
-    const allKeys = identifier.keys.map((key) => ({
-      id: identifier.did + '#' + key.kid,
-      type: keyMapping[key.type],
-      controller: identifier.did,
-      publicKeyHex: key.publicKeyHex,
-    }))
+    const contexts = new Set<string>()
+    const allKeys: VerificationMethod[] = identifier.keys.map((key: IKey) => {
+      const vm: VerificationMethod = {
+        id: identifier.did + '#' + key.kid,
+        type: keyMapping[key.type],
+        controller: identifier.did,
+        publicKeyHex: key.publicKeyHex,
+      }
+      switch (vm.type) {
+        case 'EcdsaSecp256k1VerificationKey2019':
+        case 'EcdsaSecp256k1RecoveryMethod2020':
+          contexts.add('https://w3id.org/security/v2')
+          contexts.add('https://w3id.org/security/suites/secp256k1recovery-2020/v2')
+          break
+        case 'Ed25519VerificationKey2018':
+          contexts.add('https://w3id.org/security/suites/ed25519-2018/v1')
+          vm.publicKeyBase58 = bytesToBase58(hexToBytes(key.publicKeyHex))
+          delete(vm.publicKeyHex)
+          break
+        case 'X25519KeyAgreementKey2019':
+          contexts.add('https://w3id.org/security/suites/x25519-2019/v1')
+          vm.publicKeyBase58 = bytesToBase58(hexToBytes(key.publicKeyHex))
+          delete(vm.publicKeyHex)
+          break
+        case 'Ed25519VerificationKey2020':
+          contexts.add('https://w3id.org/security/suites/ed25519-2020/v1')
+          vm.publicKeyMultibase = bytesToMultibase(hexToBytes(key.publicKeyHex), 'Ed25519')
+          delete(vm.publicKeyHex)
+          break
+        case 'X25519KeyAgreementKey2020':
+          contexts.add('https://w3id.org/security/suites/x25519-2020/v1')
+          vm.publicKeyMultibase = bytesToMultibase(hexToBytes(key.publicKeyHex), 'Ed25519')
+          delete(vm.publicKeyHex)
+          break
+        case 'EcdsaSecp256r1VerificationKey2019':
+          contexts.add('https://w3id.org/security/v2')
+          break
+        case 'Bls12381G1Key2020':
+        case 'Bls12381G2Key2020':
+          contexts.add('https://w3id.org/security/bbs/v1')
+          break
+        default:
+          break
+      }
+      return vm
+    })
     // ed25519 keys can also be converted to x25519 for key agreement
     const keyAgreementKeyIds = allKeys
       .filter((key) => ['Ed25519VerificationKey2018', 'X25519KeyAgreementKey2019'].includes(key.type))
@@ -56,7 +97,7 @@ export const WebDidDocRouter = (options: WebDidDocRouterOptions): Router => {
       .map((key) => key.id)
 
     const didDoc = {
-      '@context': 'https://w3id.org/did/v1',
+      '@context': ['https://www.w3.org/ns/did/v1', ...contexts],
       id: identifier.did,
       verificationMethod: allKeys,
       authentication: signingKeyIds,